CAMMI - Team/Assignment Task 1/requirements.txt
requests
python-docx
groq
langchain-groq
python-dotenv